Add option in snapshot_propose to rename a ProjectName
This change adds a new option:

 --current-snapshot-rename-project NAME

to snapshot-propose and snapshot-update.

It is designed to be used when you want to migrate the project name
while leaving the rest of the infrastructure untouched.  This may
be necessary when migrating a prod account.

It was tested by running it on a handful of accounts and checking
the results of shapshot-propose with diff.  The snapshot-update
script is simply a pass-through, and was tested on the ESDK
project to update the name of the project.